Split sparrow cli into multiple files
